partially implemented images upload

+++ database/migrations/2023_08_07_114847_add_phone_type_to_employee_phones_table.php
@@ -8,9 +8,11 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::table('employee_phones', function (Blueprint $table) {
-            $table->foreignId('phone_type_id')->references('id')->on('employee_phone_types')->onDelete('NO ACTION')->default(1);
-        });
+        if (!Schema::hasColumn('employee_phones', 'phone_type_id')) {
+            Schema::table('employee_phones', function (Blueprint $table) {
+                $table->foreignId('phone_type_id')->references('id')->on('employee_phone_types')->onDelete('NO ACTION')->default(1);
+            });
+        }
     }
 
     public function down()
